+++ Changelog.md
@@ -25,6 +25,7 @@ Rust:
 - in FBS, rename some field names in gates
 - in FBS, replace WireList by [WireRange] and remove type_id from WireRange
 - add conversion declarations in Relation
+- split public/private inputs files into one file per type
 
 # SIEVE IR v1.0.1, 2022-08